Fix a bug preventing key schemas being added to GSI
The `key_schema` field when creating a Global Secondary Index was
being added as an empty array. This was because originally I had
enforced `hash` type keys, and did not allow `range` keys.

The example in the last [PR](https://github.com/victorquinn/dynasty/pull/55)
included the intented use for adding both `hash` and `range`
keys, but the code for allowing Dynasty to respond to this change
was not commited.

The pull request fixes this.